requirements.txt
python-dotenv>=1.0.0
google-generativeai>=0.3.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
